// sdk/cards/src/model/enums/TemplateType.ts
/**
 * Supported template types.
 * 
 * @author Abhishek Kumar
 * @since 1.0.0
 */
enum TemplateType {

    /**
     * Basic Template
     * @since 1.0.0
     */
    BASIC = "BASIC",

    /**
     * Illustration Template
     * @since 1.0.0
     */
    ILLUSTRATION = "ILLUSTRATION"
}

export default TemplateType;