// sdk/cards/src/model/enums/WidgetType.ts
/**
 * Types of UI widgets supported.
 * 
 * @author Abhishek Kumar
 * @since 1.0.0
 */
enum WidgetType {

    /**
     * Widget that loads an image or gif
     * @since 1.0.0
     */
    IMAGE = "IMAGE",

    /**
     * Widget that loads text content.
     * @since 1.0.0
     */
    TEXT = "TEXT",

    /**
     * Widget that loads button content.
     * @since 1.0.0
     */
    BUTTON = "BUTTON"
}

export default WidgetType;